Update README.md
Hi! 
I added Teamflect to the list. It's a UK based software company that hires talent worldwide, and I work there too. I'm based in Turkiye for example. We're currently looking for new talent especially product people. So it'd be appreciated if you accept the changes. I believe we can increase our reach to connect with talented product people via this list. 

Signed-off-by: ozge113 <ozge@teamflect.com>